+++ gnssrefl/rinex3_snr.py
@@ -81,8 +81,10 @@ def main():
         rate = '30'; dec_rate = 0; archive = 'unavco' ; fortran = False; translator = 'hybrid'
         year_list = [iyear]; doy_list = [idoy]; rate = 'low';   nol = True; overwrite = False; srate = 30; mk = False; skipit = 1
         strip = False; stream = 'S'  ; bkg = 'IGS' # many of these are fake values because the file has already been translated to rinex2
+        gzip = True
+        screenstats = False
         r.run_rinex2snr(station, year_list, doy_list, isnr, orbtype, rate,dec_rate,archive,fortran,nol,
-                overwrite,translator,srate,mk,skipit,stream,strip,bkg)
+                overwrite,translator,srate,mk,skipit,stream,strip,bkg,screenstats,gzip)
 
 
 if __name__ == "__main__":